Avoid using expensive IO when there is no need to. (flushing to disk)
Avoid other IO when simply loading the game library.
This also improves booting disc games.
May or may not speed up loading the emulator with NAS or HDD drives.