@elad335 2023-04-21
Added more optimizations.