@Asinin3 2023-04-21
Before PR performance was awful. Here's my results **before** this PR with over 650 games spread across two mechanical HDDs.
0s to 10s - waiting for RPCS3 window to open
10s to 15s - waiting for Window elements to not be white
15s to 19s - waiting for gamelist to appear
19s to 27s - waiting for gamelist icons to appear
27s to 1m50s - waiting for UI to be responsive so i can actually click things
1m50s to 2m44s - waiting for disk space to be generated for all games in list

After: 
0s to 9s - waiting for RPCS3 window to open
9s to 15s - waiting for Window elements to not be white
15s to 22s - waiting for gamelist to appear
22s to 22s - waiting for gamelist icons to appear (now instant)
22s to 1m13s - waiting for UI to be responsive so i can actually click things
1m13s to 1m48s - Waiting for disk space to be generated for all games in list

So about 1minute faster for me from a cold boot. For it to finish, but more impressively 40seconds faster to actually have a responsive UI that I can click settings and things on while I wait for filesizes to still generate.

What I really want is a way to cache the file sizes RPCS3 generates for games, its the hugest hit to opening times but this is a good change