@13xforever 2023-04-21
For most people the impact won't be as big, I have ~150 games and the startup time reduced from 8s down to 2s (which is still noticeable, but not as mindblowing).