@ThePapaw 2023-04-22
I have 800+ games on NAS using virtual. I  just updated to 0.0.27-14919-6bf77166 and things are back to normal for me. I can launch RPCS3 and I can click on config and it opens right up, no more 6-8 minutes of waiting just for the config screen to open. Everything I've tried so far is back to normal for me. AWESOME work!